Refactor merge conan and container Auth preserve Actions taskID
* Remove duplicated code
* Allow further ActionsUser package permission checks
